webkit-efl/LayoutTests/platform/efl-slp/media/video-single-valid-source-expected.txt

CONSOLE MESSAGE: line 27: TypeError: 'undefined' is not a function (evaluating 'element.canPlayType(codecs[i][0])')
FAIL: Timed out waiting for notifyDone to be called
Test that a single valid <source> element loads correctly
